+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [6.8.0 - 2020-08-18]
+# [6.8.0 - 2020-09-07]
 ### Added
-- CheckScheduledInquiryJob
+- Report job config
+
+## [6.5.0 - 2020-08-10]
+### Added
+- Add tests inquiry for CQL
 
 ## [5.8.0 - 2020-05-26]
 ### Added